+++ airbyte-commons/src/main/java/io/airbyte/commons/concurrency/LifecycledCallable.java
@@ -27,22 +27,22 @@ public class LifecycledCallable<T> implements Callable<T> {
       this.onFinish = () -> {};
     }
 
-    public Builder<T> setOnStart(VoidCallable onStart) {
+    public Builder<T> setOnStart(final VoidCallable onStart) {
       this.onStart = onStart;
       return this;
     }
 
-    public Builder<T> setOnSuccess(CheckedConsumer<T, Exception> onSuccess) {
+    public Builder<T> setOnSuccess(final CheckedConsumer<T, Exception> onSuccess) {
       this.onSuccess = onSuccess;
       return this;
     }
 
-    public Builder<T> setOnException(CheckedConsumer<Exception, Exception> onException) {
+    public Builder<T> setOnException(final CheckedConsumer<Exception, Exception> onException) {
       this.onException = onException;
       return this;
     }
 
-    public Builder<T> setOnFinish(VoidCallable onFinish) {
+    public Builder<T> setOnFinish(final VoidCallable onFinish) {
       this.onFinish = onFinish;
       return this;
     }
@@ -78,7 +78,7 @@ public class LifecycledCallable<T> implements Callable<T> {
       final T result = execute();
       onSuccess(result);
       return result;
-    } catch (Exception e) {
+    } catch (final Exception e) {
       onException(e);
       throw e;
     } finally {
@@ -95,11 +95,11 @@ public class LifecycledCallable<T> implements Callable<T> {
 
   }
 
-  private void onSuccess(T value) throws Exception {
+  private void onSuccess(final T value) throws Exception {
     this.onSuccess.accept(value);
   }
 
-  private void onException(Exception e) throws Exception {
+  private void onException(final Exception e) throws Exception {
     this.onException.accept(e);
   }