+++ airbyte-integrations/bases/base-normalization/unit_tests/test_transform_config.py
@@ -134,6 +134,7 @@ class TestTransformConfig:
             "project_id": "my_project_id",
             "dataset_id": "my_dataset_id",
             "credentials_json": '{ "type": "service_account-json" }',
+            "transformation_priority": "interactive",
             "dataset_location": "EU",
         }
 
@@ -143,6 +144,7 @@ class TestTransformConfig:
             "method": "service-account-json",
             "project": "my_project_id",
             "dataset": "my_dataset_id",
+            "priority": "interactive",
             "keyfile_json": {"type": "service_account-json"},
             "location": "EU",
             "retries": 1,
@@ -164,6 +166,7 @@ class TestTransformConfig:
             "method": "oauth",
             "project": "my_project_id",
             "dataset": "my_dataset_id",
+            "priority": "interactive",
             "retries": 1,
             "threads": 32,
         }
@@ -277,6 +280,32 @@ class TestTransformConfig:
             "schema": "my_db",
             "database": "my_db",
             "username": "a user",
+            "password": "password1234",
+        }
+
+        assert expected == actual
+        # DBT schema is equivalent to MySQL database
+        assert extract_schema(actual) == "my_db"
+
+    def test_transform_mssql(self):
+        input = {
+            "type": "sqlserver",
+            "host": "airbyte.io",
+            "port": 1433,
+            "database": "my_db",
+            "schema": "my_db",
+            "username": "SA",
+            "password": "password1234",
+        }
+
+        actual = TransformConfig().transform_mysql(input)
+        expected = {
+            "type": "sqlserver",
+            "server": "airbyte.io",
+            "port": 1433,
+            "schema": "my_db",
+            "database": "my_db",
+            "username": "SA",
             "password": "password1234",
         }