+++ airbyte-integrations/connectors/source-mssql/src/main/java/io/airbyte/integrations/source/mssql/MssqlCdcTargetPosition.java
@@ -23,24 +23,24 @@ public class MssqlCdcTargetPosition implements CdcTargetPosition {
   private static final Logger LOGGER = LoggerFactory.getLogger(MssqlCdcTargetPosition.class);
   public final Lsn targetLsn;
 
-  public MssqlCdcTargetPosition(Lsn targetLsn) {
+  public MssqlCdcTargetPosition(final Lsn targetLsn) {
     this.targetLsn = targetLsn;
   }
 
   @Override
-  public boolean reachedTargetPosition(JsonNode valueAsJson) {
-    Lsn recordLsn = extractLsn(valueAsJson);
+  public boolean reachedTargetPosition(final JsonNode valueAsJson) {
+    final Lsn recordLsn = extractLsn(valueAsJson);
 
     if (targetLsn.compareTo(recordLsn) > 0) {
       return false;
     } else {
-      SnapshotMetadata snapshotMetadata = SnapshotMetadata.valueOf(valueAsJson.get("source").get("snapshot").asText().toUpperCase());
+      final SnapshotMetadata snapshotMetadata = SnapshotMetadata.valueOf(valueAsJson.get("source").get("snapshot").asText().toUpperCase());
       // if not snapshot or is snapshot but last record in snapshot.
       return SnapshotMetadata.TRUE != snapshotMetadata;
     }
   }
 
-  private Lsn extractLsn(JsonNode valueAsJson) {
+  private Lsn extractLsn(final JsonNode valueAsJson) {
     return Optional.ofNullable(valueAsJson.get("source"))
         .flatMap(source -> Optional.ofNullable(source.get("commit_lsn").asText()))
         .map(Lsn::valueOf)
@@ -48,14 +48,14 @@ public class MssqlCdcTargetPosition implements CdcTargetPosition {
   }
 
   @Override
-  public boolean equals(Object o) {
+  public boolean equals(final Object o) {
     if (this == o) {
       return true;
     }
     if (o == null || getClass() != o.getClass()) {
       return false;
     }
-    MssqlCdcTargetPosition that = (MssqlCdcTargetPosition) o;
+    final MssqlCdcTargetPosition that = (MssqlCdcTargetPosition) o;
     return targetLsn.equals(that.targetLsn);
   }
 
@@ -64,21 +64,21 @@ public class MssqlCdcTargetPosition implements CdcTargetPosition {
     return targetLsn.hashCode();
   }
 
-  public static MssqlCdcTargetPosition getTargetPosition(JdbcDatabase database, String dbName) {
+  public static MssqlCdcTargetPosition getTargetPosition(final JdbcDatabase database, final String dbName) {
     try {
       final List<JsonNode> jsonNodes = database
           .bufferedResultSetQuery(conn -> conn.createStatement().executeQuery(
               "USE " + dbName + "; SELECT sys.fn_cdc_get_max_lsn() AS max_lsn;"), JdbcUtils.getDefaultSourceOperations()::rowToJson);
       Preconditions.checkState(jsonNodes.size() == 1);
       if (jsonNodes.get(0).get("max_lsn") != null) {
-        Lsn maxLsn = Lsn.valueOf(jsonNodes.get(0).get("max_lsn").binaryValue());
+        final Lsn maxLsn = Lsn.valueOf(jsonNodes.get(0).get("max_lsn").binaryValue());
         LOGGER.info("identified target lsn: " + maxLsn);
         return new MssqlCdcTargetPosition(maxLsn);
       } else {
         throw new RuntimeException("SQL returned max LSN as null, this might be because the SQL Server Agent is not running. " +
             "Please enable the Agent and try again (https://docs.microsoft.com/en-us/sql/ssms/agent/start-stop-or-pause-the-sql-server-agent-service?view=sql-server-ver15)");
       }
-    } catch (SQLException | IOException e) {
+    } catch (final SQLException | IOException e) {
       throw new RuntimeException(e);
     }
   }