+++ airbyte-integrations/connectors/source-snowflake/src/test-integration/java/io/airbyte/integrations/io/airbyte/integration_tests/sources/SnowflakeSourceAcceptanceTest.java
@@ -93,7 +93,7 @@ public class SnowflakeSourceAcceptanceTest extends SourceAcceptanceTest {
 
   // for each test we create a new schema in the database. run the test in there and then remove it.
   @Override
-  protected void setupEnvironment(TestDestinationEnv environment) throws Exception {
+  protected void setupEnvironment(final TestDestinationEnv environment) throws Exception {
     config = Jsons.clone(getStaticConfig());
     database = Databases.createJdbcDatabase(
         config.get("username").asText(),
@@ -128,7 +128,7 @@ public class SnowflakeSourceAcceptanceTest extends SourceAcceptanceTest {
   }
 
   @Override
-  protected void tearDown(TestDestinationEnv testEnv) throws Exception {
+  protected void tearDown(final TestDestinationEnv testEnv) throws Exception {
     final String dropSchemaQuery = String
         .format("DROP SCHEMA IF EXISTS %s", SCHEMA_NAME);
     database.execute(dropSchemaQuery);