+++ airbyte-integrations/connectors/source-zendesk-support/integration_tests/integration_test.py
@@ -7,7 +7,7 @@ import json
 import pendulum
 import requests_mock
 from source_zendesk_support import SourceZendeskSupport
-from source_zendesk_support.streams import Macros, TicketAudits, TicketMetrics, Tickets, Users
+from source_zendesk_support.streams import LAST_END_TIME_KEY, Macros, TicketAudits, TicketMetrics, Tickets, Users
 
 CONFIG_FILE = "secrets/config.json"
 
@@ -32,12 +32,12 @@ class TestIntegrationZendeskSupport:
             # save the first 5 records
             if len(record_timestamps) > 5:
                 break
-            if stream._last_end_time not in record_timestamps.values():
-                record_timestamps[record["id"]] = stream._last_end_time
+            if stream.last_end_time not in record_timestamps.values():
+                record_timestamps[record["id"]] = stream.last_end_time
 
         stream.page_size = 10
         for record_id, timestamp in record_timestamps.items():
-            state = {"_last_end_time": timestamp}
+            state = {LAST_END_TIME_KEY: timestamp}
             for record in stream.read_records(sync_mode=None, stream_state=state):
                 assert record["id"] != record_id
                 break