+++ airbyte-integrations/bases/source-acceptance-test/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.1.24
+Improve message about errors in the stream's schema: https://github.com/airbytehq/airbyte/pull/6934
+
+## 0.1.23
+Fix incorrect auth init flow check defect.
+
 ## 0.1.22
 Fix checking schemas with root $ref keyword